+++ app/src/main/java/com/hldj/hmyg/saler/purchase/PurchaseSeedlingParmAdapter.java
@@ -44,7 +44,7 @@ public class PurchaseSeedlingParmAdapter extends BaseAdapter {
 		l_params = new LinearLayout.LayoutParams(LayoutParams.MATCH_PARENT,
 				LayoutParams.WRAP_CONTENT);
 		WindowManager wm = ((Activity) context).getWindowManager();
-		l_params.height = (int) (wm.getDefaultDisplay().getWidth() * 9 / 20);
+		l_params.height = wm.getDefaultDisplay().getWidth() * 9 / 20;
 	}
 
 	@Override
